Added Idle and Walk animation for Player
The PlayerAnimator subscribes to a PlayerWalking event and is notified when the player is walking or not walking which controls the animation.
The animation moves the body and the head.
When the player is idle, the head moves up and down.
When the player is walking, the body and head move up and down, but faster.