Refactored Counter Classes, Added Container Counter and used Prefab Variant on the Counter
Introduced BaseCounter abstract class as the base class for all counter types. This change adds further structure to the counter classes and makes the code more maintainable and extensible. The BaseCounter class also implements the IKitchenObjectParent interface, allowing more generalized interaction with the KitchenObject.

Also created ContainerCounter and KitchenCounter classes, which inherit from BaseCounter and provide different interaction behaviors, demonstrating the benefit of this new class hierarchy.

The Player class interacts with BaseCounter objects instead of specific counter types. This allows the player to interact with any type of counter in a flexible and extensible way.

This refactoring makes the code more structured, maintainable, and ready for further expansion and addition of new features.