Implemented functionality for plate interactions and ingredient plating
In this commit, the capability of adding ingredients to a plate has been introduced. New classes PlateKitchenObject and PlatingRecipeSO have been added to handle these interactions. The KitchenCounter and CuttingCounter classes have been updated to support the new plate functionalities. Ingredients can now be added to a plate object, and a complete recipe can be formed by adding all necessary ingredients to a single plate.