Needed to reset lastMoveMade and reattachTime
User may tap the movement key multiple times and go between moving and stopping.  When this happens it short circuits on the client and needs to not reattach.  It could be in the middle of attempting to reattach so resetting these variables will move into the right state of moving.