Changed the logic for punishment and Rolling Latency
There was a bug that was using the time from the server to subtract from the accumulatedeltatime, since that delta time is coming from the client this will cause a conflict.  Instead we work around this by just cleaning up the accumulatedDeltaTime if more than 2 seconds has passed in the last input. We can safely assume that the player stopped if IsRunning is false and timeSinceLastCommand is greater than 2 seconds.
So after 30 seconds we give the player back the ability to walk at normal speed if they have not moved and they are not cheating again.