Removed Client Rpc update from ServersSoundManager
Bug fix: I needed a dictionary to deal with multiple players, I didnt understand that during the ServerRpc that there was only 1 instance and not multiple instances on the server.
Refactor: Single responsibility principle, created a ClientRpcManager and moved the functionality into there.
Updated all of the Kitchen Objects to NetworkObjects so that the Server can instantiate them over the network.