Implement Network Enhancements and Optimize ServerRPC Communication
Integrated multiple features and workarounds to facilitate network synchronization and efficient ServerRpc communications:

- Enhanced kitchen-related classes (`BaseCounter.cs`, `IKitchenObjectParent.cs`, and `KitchenObject.cs`) to work with Unity's Netcode.
- Introduced `FollowTransform.cs` to allow KitchenObjects to dynamically follow players, addressing the challenge of direct NetworkObject parenting.
- Developed `KitchenObjectListSO` as a solution to the limitations of sending ScriptableObjects via ServerRpc. Employed an index-based lookup system to simplify kitchen object communication.
- Integrated the index-based system in `KitchenGameMultiplayer.cs` for streamlined ServerRpc interactions.
KitchenGameMultiplayer.cs: Manages the spawning and synchronization of kitchen objects across multiple clients. This file plays a crucial role in ensuring that all players see consistent states of kitchen objects in a multiplayer environment, especially when objects are picked up, placed, or interacted with. It leverages the KitchenObjectListSO for efficient ServerRpc communications, utilizing an index-based lookup system due to the limitations of sending ScriptableObjects directly via ServerRpc.