Bug Fix: Visibility when laggy
If there was lag the visibility on the client could be set locally twice before the network had set the visibility once.
What this means is that the client may dispose of a kitchen object which would mean the item should be hidden but then the network will sync and set the visibility to true when it shouldn't have because the items has been disposed of.  To fix this I added a visibilitySetLocally counter and when the network finally gets around to updating the visibility I will not set the value until the counter has reached zero.