Estimate the InitialSize and TimeoutMilliseconds with the latency
When deciding how many items to spawn ahead of time in the object pooler, allow for 10 items to be requested, assuming a user can grab an item every 3 seconds. If the user is fast enough and requests a 11th item and they have poor lag, they may need to wait for the item to appear.